Split monitor mode out into a separate tool
This makes for a cleaner separation of concerns. The anystat binary collects
data and writes it out to log-files, a database or a socket; the monitor
binary can be run on the database to show graphs and statistics. This makes it
easier to run anystat as a daemon and avoids the resource usage of the graphs
when no-one is logged in anyway. For now the functionality of the monitor is a
bit more limited than the monitoring mode used to be, but that will be filled
in over time.